got the blue screen to render :)

+++ eucalyptus/src/main.rs
@@ -1,24 +1,25 @@
 mod scene1;
-use dropbear_engine::{input, WindowConfiguration};
+use dropbear_engine::{WindowConfiguration, input};
 
 fn main() {
     let config = WindowConfiguration {
-        width: 1280 as u32,
-        height: 720 as u32,
-        title: "Eucalyptus, built with dropbear"
+        width: 1280u32,
+        height: 720u32,
+        title: "Eucalyptus, built with dropbear",
     };
 
     let _app = dropbear_engine::run_app!(config, |scene_manager, input_manager| {
         let testing_scene = scene1::TestingScene1::new();
         let testing_keyboard = scene1::TestingScene1::new();
         let testing_mouse = scene1::TestingScene1::new();
-        
+
         scene_manager.add("testing_scene", Box::new(testing_scene));
         input_manager.add_keyboard("testing_scene_keyboard", Box::new(testing_keyboard));
         input_manager.add_mouse("testing_scene_mouse", Box::new(testing_mouse));
-        
+
         scene_manager.attach_input("testing_scene", "testing_scene_keyboard");
         scene_manager.attach_input("testing_scene", "testing_scene_mouse");
         scene_manager.switch("testing_scene");
-    }).unwrap();
+    })
+    .unwrap();
 }
